Move nonnull type to public namespace
